修复：清理临时分享 alist：获取分享链接失败自动重试下一个帐号

+++ src/main/java/cn/har01d/alist_tvbox/service/ShareService.java
@@ -140,7 +140,7 @@ public class ShareService {
         pikPakService.readPikPak();
         driverAccountService.loadStorages();
 
-        cleanTempShares();
+        cleanTempShares(false);
 
         List<Share> list = shareRepository.findAll();
         fixPath(list);
@@ -209,23 +209,25 @@ public class ShareService {
 
     @Scheduled(cron = "0 30 * * * *")
     public void cleanShares() {
-        cleanTempShares();
+        cleanTempShares(true);
         cleanInvalidShares();
     }
 
-    private void cleanTempShares() {
+    private void cleanTempShares(boolean delete) {
         List<Share> list = shareRepository.findByTempTrue();
         Instant time = Instant.now().minus(appProperties.getTempShareExpiration(), ChronoUnit.HOURS);
+        String token = accountService.login();
         for (Share share : list) {
             if (share.isTemp() && share.getTime() != null && share.getTime().isBefore(time)) {
                 log.info("Delete temp share: {} {}", share.getId(), share.getPath());
                 shareRepository.delete(share);
 
-                try {
-                    String token = accountService.login();
-                    deleteStorage(share.getId(), token);
-                } catch (Exception e) {
-                    log.warn("cleanTempShare error", e);
+                if (delete) {
+                    try {
+                        deleteStorage(share.getId(), token);
+                    } catch (Exception e) {
+                        log.warn("cleanTempShare error", e);
+                    }
                 }
             }
         }
